+++ src/main/java/sardorcreate/dto/ip/IpDto.java
@@ -5,6 +5,7 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.ToolsStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -18,6 +19,6 @@ public class IpDto {
     private String model;
     private LocalDate date;
     private String whereFrom;
-    private long price;
+    private BigDecimal price;
     private ToolsStatus status;
 }
