+++ src/main/java/sardorcreate/dto/printer/PrinterCreateDto.java
@@ -5,6 +5,8 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.PrinterType;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -13,7 +15,7 @@ public class PrinterCreateDto {
     private long inventoryId;
     private String model;
     private String whereFrom;
-    private long price;
+    private BigDecimal price;
     private PrinterType type;
     private boolean isColored;
     private int colorCount;
