+++ src/main/java/sardorcreate/dto/printer/PrinterDto.java
@@ -6,6 +6,7 @@ import lombok.ToString;
 import sardorcreate.enums.PrinterType;
 import sardorcreate.enums.ToolsStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -19,7 +20,7 @@ public class PrinterDto {
     private String model;
     private LocalDate date;
     private String whereFrom;
-    private long price;
+    private BigDecimal price;
     private ToolsStatus status;
     private PrinterType type;
     private boolean isColored;
