+++ src/main/java/sardorcreate/dto/sep_monitor/SepMonitorCreateDto.java
@@ -5,6 +5,8 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.ScreenType;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -13,7 +15,7 @@ public class SepMonitorCreateDto {
     private long inventoryId;
     private String model;
     private String whereFrom;
-    private long price;
+    private BigDecimal price;
     private int size;
     private int refreshRate;
     private ScreenType monType;
