+++ src/main/java/sardorcreate/dto/tablet/TabletCreateDto.java
@@ -5,6 +5,8 @@ import lombok.Setter;
 import lombok.ToString;
 import sardorcreate.enums.OSType;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -13,7 +15,7 @@ public class TabletCreateDto {
     private long inventoryId;
     private String model;
     private String whereFrom;
-    private long price;
+    private BigDecimal price;
     private OSType osType;
     private int ram;
     private int rom;
